Add PyPy support by making sure to close all the files we write to.
Because PyPy doesn't use a refcounted GC, leaving dangling files open
causes the writes to be flushed at some non-deterministic time in the
future, meaning that prerelease (and anything that wanted to write to
the users files) typically wound up committing an empty file.

Most of the fixes were in the tests but there were several in the
actual code itself. I didn't fix all the read() cases in the tests for
clarity. Those do leave dangling files and so running out of
file-descriptors is a possibility, but not an immediate practical concern.

Note: The svn.txt test fails locally; I think it's just due to a
difference in the output by different SVN versions.