Always exit with error code 1 when we exit explicitly.
In some cases we would exit with success code 0 when we exited based
on the answer to a question.  This happened when the user did not want
us to create the missing tags directory in subversion, and also after
asking if the user wanted to continue even though 'nothing changed
yet' was in the history.