Allow `pypirc` with just a `pypi` section.
Previously, we required either a `server-login` section with a
`username` option, or a `distutils` section with an `index-servers`
option.  Failing this, we gave a warning about a not properly
configured file, and happily continued without uploading anything.
Now if there is something missing from the `pypirc` file, we give an
error and explicitly ask if you want to continue without uploading.

Fixes issue #199.

Note for developers of extensions for `zest.releaser`: this removes
the `is_old_pypi_config` and `is_new_pypi_config` methods, because
they made no sense anymore.  If you were using these, see if you can
use the `distutils_server` method instead.