Refactored data documentation for entry points.
The common data that is shared by all or most commands, has been moved
to baserelease.py.  For their data dicts they now use a copy of this.

The common data items are now listed separately in the entry points
documentation.  For the commands we only show the data that is
different: extra keys, or keys with a different value (help text).

Note that not all common keys are available for all commands, or not
in all entry points because they may be calculated later.  But we were
already warning about this in the documentation.