Subversion fix: create tag of entire trunk or branch when not in repo root.
If you have ``trunk/pkg1`` and ``trunk/pkg2`` and you make tag 1.0 in directory ``pkg1``,
then until now we would create ``tags/1.0`` with the contents of directory ``pkg1``.
Checking out the tag and changing to the ``pkg1`` directory then failed.
We now make a tag of the entire trunk or branch, just like in the other version control systems.

Fixes issue #213.