Fixed more ugly printing.
Printing the twine command for retrying would give this:

    Command failed: "t w i n e ' ' u p l o a d"

And with Python2.7, in this and various other spots, a ``u'unicode'`` marker would be visible.
This is because several things were logged with ``%r``.
In most cases, the only reason for the raw format was so that it was clear where the command started and ended,
because it would be enclosed in 'quotes'.
But for unicode this becomes u'quotes', which is ugly.
So instead of `%r` we now quote it ourselves with `'%s'`.