Added keys update_history and update_version in data.
This is in prerelease and postrelease data.
Plugins can use this to tell `zest.releaser` (and other plugins) to not touch the history or the version number.
Default value is True: do update history and version.

The `update_history` key will be used by a towncrier plugin that will update the changelog with news fragments.

I do not have a use for the `update_version` key yet, but it seemed a logical data key to add.

The towncrier plugin will add an internal data key `_towncrier_applicable`.
With this, you would get this warning when running zest.releaser:

    Internal detail: key(s) ['_towncrier_applicable'] are not documented

So I have changed this internal check to allow undocumented data keys when they start with underscore.