Added key update_history in data.
This is in prerelease and postrelease data.
Plugins can use this to tell `zest.releaser` (and other plugins) to not touch the history, presumably because the plugin itself handles it.

Default value is True: do update history.

The `update_history` key will be used by a towncrier plugin that will update the changelog with news fragments.

The towncrier plugin will add an internal data key `_towncrier_applicable`.
With this, you would get this warning when running zest.releaser:

    Internal detail: key(s) ['_towncrier_applicable'] are not documented

So I have changed this internal check to allow undocumented data keys when they start with underscore.