Add towncrier hooks for updating the changelog with news fragments.
- Before prerelease, check that towncrier is available and configured.
  Set a marker to inform zest.releaser to not update the history.

- In the middle of prerelease, call towncrier to update the changelog with newsfragments.

- Before postrelease, again check that towncrier is available and configured.
  Set a marker to inform zest.releaser to not update the history.
  Towncrier itself does not need to do anything in this phase.

The towncrier hooks set `update_history` to False in the prerelease/postrelease data  when towncrier should be run.
The towncrier hooks also set `_towncrier_applicable` True/False in the data.
This avoids running the towncrier checks multiple times.
We could look for `update_history = False` instead, but theoretically in the future some other hook may be setting this.
So it seems good to split this.

We may want to split the above part into a separate package.
But it needs a few minor additions to zest.releaser, which are in the previous commit.