Start removal of svn/hg/bzr from rest of files.
This also meant using Git instead of Subversion in for example `prerelease.txt`.
This meant I needed to add more explicit commits of test changes, otherwise git complained about a non-clean checkout.
So I created a test function `commit_all_changes` for this.

Also, I added a few `git config` commands in the setup in `functional.py`.
I set a local user name and email, the same as what we have in Dockerfile, so it can perhaps be removed there.

And I locally disabled signing of commits and tags.
I have *enabled* this in my own `~/.gitconfig` and this interfered with the tests, giving lots of errors, also before my big refactoring.