Use native namespace packages for zest.releaser.
This is instead of deprecated pkg_resources based ones.

Note that this can be a problem in case `zest.releaser` is installed in combination with other packages in the `zest` namespace.
But there are only a few such Plone related packages, so they won't usually be installed in the same virtual environment.
And otherwise I can make new releases of those packages.
Now seems a good time to do this.