When a command we call exits with a non-zero exit code, clearly state this in the output.
Note that this is tricky to do right.  Some commands like ``git`` seem to print everything to stderr,
leading us to think there are errors, but the exit code is zero, so it should be fine.
We do not want to ask the user to often if she wants to continue or not.
But we do not want to silently swallow important errors either.