When a command fails, always ask the user to continue or not.
Interestingly, this needed some fixes in the tests, where commands had gone wrong for years but we never noticed.

* At the end of `fullrelease` we try to push, and this always failed because there was no push target.
  We now ignore the push commands and instead print them.
* One test renames `CHANGES.rst` to `CHANGES.md`, and then calling `python setup.py --version` actually fails:
  `setup.py` still tries to read the no longer existing `CHANGES.rst`.

I hope we now strike a good balance between showing error output, exit codes, and asking if the user wants to continue.
With the previous alpha release, the balance was not good for me.  The final `git push` used `stderr` to report that everything was fine,
and then `postrelease` treated this as an error, and asked me if I wanted to continue.  Now we check the exit code.