feat(prerelease): append changes on commit message
Grab the changes that are to be written on the changelog file and write
them as part of the commit message.

This allows to integrate with GitHub release tooling to automatically
grab what's part of a release from the commit message of said release.

This way, while still releasing with `zest.releser` one can nicely
integrate with GitHub releases (and probably GitLab releases as well).