Fix tests: with setuptools>=69.3 sdists get an underscore.
Instead of `tha.example-0.1.tar.gz` you get `tha_example-0.1.tar.gz`.
Wheels currently still have dots there, but those might become underscores at some point as well.

This is actually a problem for Buildout: https://github.com/buildout/buildout/issues/647

The change in this PR does have a bit of an effect:

- Previously when we were running `build`, it would run in non-isolated mode.  So if you make sure you have an older `setuptools` version in the venv of `zest.releaser`, you get dots, like was always the case.
- Now with this PR, we run in isolated mode, which means that `build` will most likely install the latest setuptools version.  Well, it does that if setuptools is the chosen build-system, and I think this is still the default, at least if there is a `setup.py`.

You can influence this in the `pyproject.toml` of the package you are releasing.
In Plone we have this in most packages:

```
[build-system]
requires = ["setuptools>=68.2"]
```

I think we want it like this now:

```
[build-system]
requires = ["setuptools>=68.2,<=69.2"]
```

I am mostly writing this as note to myself. :-)