Fix bumpversion --final.
When a version has no alpha/beta/rc, we want the suggestion to be None.

In the tests we were checking `utils.suggest_version('6.0.0dev0', final=True)`, which indeed gives None.
But that is not how the version normally looks.  A dot is missing.  And with the dot it no longer does what we want:
`utils.suggest_version('6.0.0.dev0', final=True)` gives `6.0.1.dev0` instead of None.