Improvements for bumpversion (#478)
I squashed the commits from https://github.com/zestsoftware/zest.releaser/pull/478 into one while merging. Otherwise there are too many changes that overwrite each other.

* Fix bumpversion --final.

When a version has no alpha/beta/rc, we want the suggestion to be None.

In the tests we were checking `utils.suggest_version('6.0.0dev0', final=True)`, which indeed gives None.
But that is not how the version normally looks.  A dot is missing.  And with the dot it no longer does what we want:
`utils.suggest_version('6.0.0.dev0', final=True)` gives `6.0.1.dev0` instead of None.

* Update pre-commit-hooks version.

I saw this warning when making a commit:

```
[INFO] Initializing environment for https://github.com/pre-commit/pre-commit-hooks.
[WARNING] repo `https://github.com/pre-commit/pre-commit-hooks` uses deprecated stage names (commit, push) which will be removed in a future version.  Hint: often `pre-commit autoupdate --repo https://github.com/pre-commit/pre-commit-hooks` will fix this.  if it does not -- consider reporting an issue to that repo.
```

* bumpversion command: added --alpha, --beta, and --rc options.

Use this to set the version to an alpha, beta, or release candidate version.

* Allow combining ``--alpha/--beta/--rc`` with ``--feature/--breaking``.

Other combinations are still mutually exclusive.

* Document the new bumpversion data.

* Fix bumpversion tests.

Calling `--final` on a version `3.0.0.dev0` no longer does anything.
Previously it would set the version to `3.0.1.dev0`, even though `3.0.0` was not released yet.

* Actually add alpha/beta/rc options to bumpversion.

I had only added them to utils.suggest_version so far.

* Add alpha/beta/rc options to postrelease command.

* Let bumpversion be smart about combinations of release and prerelease.

* Do not support breaking/feature release on a pre release.

* Bumped version for feature release.

* Remove no longer needed commented out code.

* Add utils.check_mutially_exclusive_options function.

We use this in both bumpversion and postrelease.

* Apply suggestion from @reinout for readme

Co-authored-by: Reinout van Rees <reinout@vanrees.org>

---------

Co-authored-by: Reinout van Rees <reinout@vanrees.org>